Helpers for creating community facing hubverse package documents.

These functions create community facing documents detailing how contributors and the wider community can interact with hubverse packages. Each function creates the relevant file in the .github directory from templates and is set up with appropriate hubverse default values.

Usage

use_hubdev_coc(
  contacts =
    "Lucie Contamin (contamin@pitt.edu),\n                           Emily Howerton (ehowerton@psu.edu) or\n                           Seb Funk (sebastian.funk@lshtm.ac.uk)"
)

use_hubdev_contributing(
  organisation = "Infectious-Disease-Modeling-Hubs",
  hubdocs_contribute_url =
    "https://hubdocs.readthedocs.io/en/latest/overview/contribute.html"
)

Arguments

contacts

Character string. Contact name and email of hubverse community members responsible for enforcing the Code of Conduct.

organisation

Character string. The hubverse GitHub organisation name. The user's role and the token's scopes must be such that you have permission to create repositories in the hubverse organisation.

hubdocs_contribute_url

Character string. URL of the general contributing information page on hubDocs.